refactor(step4): merge car movement manager class responsibility to cars class
