+++ src/main/kotlin/racingcar/RacingcarGameApplication.kt
@@ -1,13 +1,7 @@
 package racingcar
 
-import racingcar.component.CarMovementManager
-
 fun main() {
-    val carMovementManager = CarMovementManager()
-
-    val application = RacingcarGame(
-        carMovementManager = carMovementManager
-    )
+    val application = RacingcarGame()
 
     application.play()
 }
